Refactor function names, make the wait list non-static so it can be tested, move the wait list to its own file
