+++ innative-env/innative-env.vcxproj.filters
@@ -24,12 +24,18 @@
     <ClCompile Include="atomics.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="wait_list.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="internal.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="atomics.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="wait_list.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
